Extract shared PickerButton component
